move font logic to fonts package
