+++ packages/asset-uploader/src/utils/format-asset.ts
@@ -1,22 +1,14 @@
 import { z } from "zod";
 import { Asset as DbAsset } from "@webstudio-is/prisma-client";
-import { type FontFormat, FONT_FORMATS } from "@webstudio-is/fonts";
+import { type FontFormat, FONT_FORMATS, FontMeta } from "@webstudio-is/fonts";
 import { Asset } from "../types";
 import { getAssetPath } from "./get-asset-path";
-import { styles } from "./get-font-data";
 
 const ImageMeta = z.object({
   width: z.number(),
   height: z.number(),
 });
 export type ImageMeta = z.infer<typeof ImageMeta>;
-
-export const FontMeta = z.object({
-  family: z.string(),
-  style: z.enum(styles),
-  weight: z.number(),
-});
-export type FontMeta = z.infer<typeof FontMeta>;
 
 export const formatAsset = (asset: DbAsset): Asset => {
   const base = { ...asset, path: getAssetPath(asset) };
