+++ packages/fonts/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require("@webstudio-is/jest-config");
