+++ packages/fonts/package.json
@@ -6,17 +6,22 @@
   "homepage": "https://webstudio.is",
   "scripts": {
     "typecheck": "tsc --noEmit",
+    "test": "jest",
     "checks": "yarn typecheck && yarn lint",
     "dev": "tsup src/index.ts --format esm,cjs --watch --out-dir=lib",
     "build": "rm -fr lib && tsc",
     "lint": "eslint ./src --ext .ts,.tsx --max-warnings 0"
   },
+  "dependencies": {
+    "zod": "^3.19.1",
+    "fontkit": "^2.0.2"
+  },
   "devDependencies": {
-    "@webstudio-is/asset-uploader": "*",
-    "@webstudio-is/jest-config": "*",
     "@webstudio-is/design-system": "*",
+    "@webstudio-is/jest-config": "*",
     "tsup": "^6.1.3",
-    "typescript": "4.7.4"
+    "typescript": "4.7.4",
+    "@types/fontkit": "^1.8.0"
   },
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
