+++ packages/fonts/src/index.ts
@@ -1,3 +1,5 @@
 export * from "./constants";
 export * from "./get-font-faces";
 export * from "./types";
+export * from "./get-font-data";
+export * from "./schema";
